+++ catscripts/Government/Countries/United_States/US_states/Alabama/Alabama.sh
@@ -3,13 +3,23 @@
 KEYWORDS_ALABAMA="Alabama"
 KEYWORDS_ALABAMA_EXCLUDE="Sweet(| )Home(| )Alabama"
 
-ALABAMA=$(egrep -i "$KEYWORDS_ALABAMA" newpages.txt | egrep -iv "$KEYWORDS_ALABAMA_EXCLUDE") 
-
-if [ "$ALABAMA" != "" ];
+if [ "$1" == "" ];
 then
-  egrep -i "$KEYWORDS_ALABAMA" newpages.txt | egrep -iv "$KEYWORDS_ALABAMA_EXCLUDE" > Alabama.txt
-  export CATFILE="Alabama.txt"
-  export CATNAME="Alabama"
-  $CATEGORIZE
-  rm Alabama.txt
+
+  debug_start "Alabama"
+
+  ALABAMA=$(egrep -i "$KEYWORDS_ALABAMA" newpages.txt | egrep -iv "$KEYWORDS_ALABAMA_EXCLUDE") 
+
+  if [ "$ALABAMA" != "" ];
+  then
+    printf "%s" "$ALABAMA" > Alabama.txt
+    export CATFILE="Alabama.txt"
+    export CATNAME="Alabama"
+    $CATEGORIZE
+    rm Alabama.txt
+    unset ALABAMA
+  fi
+
+  debug_end "Alabama"
+
 fi
