+++ catscripts/Government/Countries/United_States/US_states/Alaska/Alaska.sh
@@ -2,13 +2,23 @@
 
 KEYWORDS_ALASKA="Alaska|Seward's(| )icebox"
 
-ALASKA=$(egrep -i "$KEYWORDS_ALASKA" newpages.txt)
-
-if [ "$ALASKA" != "" ];
+if [ "$1" == "" ];
 then
-  egrep -i "$KEYWORDS_ALASKA" newpages.txt > Alaska.txt
-  export CATFILE="Alaska.txt"
-  export CATNAME="Alaska"
-  $CATEGORIZE
-  rm Alaska.txt
+
+  debug_start "Alaska"
+
+  ALASKA=$(egrep -i "$KEYWORDS_ALASKA" newpages.txt)
+
+  if [ "$ALASKA" != "" ];
+  then
+    printf "%s" "$ALASKA" > Alaska.txt
+    export CATFILE="Alaska.txt"
+    export CATNAME="Alaska"
+    $CATEGORIZE
+    rm Alaska.txt
+    unset ALASKA
+  fi
+
+  debug_end "Alaska"
+
 fi
