+++ catscripts/Government/Countries/United_States/US_states/Connecticut/Connecticut.sh
@@ -2,14 +2,23 @@
 
 KEYWORDS_CONNECTICUT="Connecticut|Hartford|, CT\b"
 
-CONNECTICUT=$(egrep -i "$KEYWORDS_CONNECTICUT" newpages.txt)
-
-if [ "$CONNECTICUT" != "" ];
+if [ "$1" == "" ];
 then
-  egrep -i "$KEYWORDS_CONNECTICUT" newpages.txt > Connecticut.txt
-  export CATFILE="Connecticut.txt"
-  export CATNAME="Connecticut"
-  $CATEGORIZE
-  rm Connecticut.txt
-  unset CONNECTICUT
+
+  debug_start "Connecticut"
+
+  CONNECTICUT=$(egrep -i "$KEYWORDS_CONNECTICUT" newpages.txt)
+
+  if [ "$CONNECTICUT" != "" ];
+  then
+    printf "%s" "$CONNECTICUT" > Connecticut.txt
+    export CATFILE="Connecticut.txt"
+    export CATNAME="Connecticut"
+    $CATEGORIZE
+    rm Connecticut.txt
+    unset CONNECTICUT
+  fi
+
+  debug_end "Connecticut"
+
 fi
