+++ catscripts/Government/Countries/United_States/US_states/Ohio/Ohio.sh
@@ -7,33 +7,45 @@ KEYWORDS_CINCINNATI="Cincinnati"
 KEYWORDS_CINCINNATI_EXCLUDE="Cincinnati(| )Bengals"
 KEYWORDS_OHIO_EXCLUDE="$KEYWORDS_CLEVELAND|$KEYWORDS_CINCINNATI"
 
-OHIO=$(egrep -i "$KEYWORDS_OHIO" newpages.txt | egrep -iv "$KEYWORDS_OHIO_EXCLUDE")
-CLEVELAND=$(egrep -i "$KEYWORDS_CLEVELAND" newpages.txt | egrep -iv "$KEYWORDS_CLEVELAND_EXCLUDE")
-CINCINNATI=$(egrep -i "$KEYWORDS_CINCINNATI" newpages.txt | egrep -iv "$KEYWORDS_CINCINNATI_EXCLUDE")
-
-if [ "$OHIO" != "" ];
+if [ "$1" == "" ];
 then
-  printf "%s" "$OHIO" > Ohio.txt
-  export CATFILE="Ohio.txt"
-  export CATNAME="Ohio"
-  $CATEGORIZE
-  rm Ohio.txt
-fi
 
-if [ "$CLEVELAND" != "" ];
-then
-  printf "%s" "$CLEVELAND" > Cleveland.txt
-  export CATFILE="Cleveland.txt"
-  export CATNAME="Cleveland"
-  $CATEGORIZE
-  rm Cleveland.txt
-fi
+  debug_start "Ohio"
 
-if [ "$CINCINNATI" != "" ];
-then
-  printf "%s" "$CINCINNATI" > Cincinnati.txt
-  export CATFILE="Cincinnati.txt"
-  export CATNAME="Cincinnati"
-  $CATEGORIZE
-  rm Cincinnati.txt
+  OHIO=$(egrep -i "$KEYWORDS_OHIO" newpages.txt | egrep -iv "$KEYWORDS_OHIO_EXCLUDE")
+  CLEVELAND=$(egrep -i "$KEYWORDS_CLEVELAND" newpages.txt | egrep -iv "$KEYWORDS_CLEVELAND_EXCLUDE")
+  CINCINNATI=$(egrep -i "$KEYWORDS_CINCINNATI" newpages.txt | egrep -iv "$KEYWORDS_CINCINNATI_EXCLUDE")
+
+  if [ "$OHIO" != "" ];
+  then
+    printf "%s" "$OHIO" > Ohio.txt
+    export CATFILE="Ohio.txt"
+    export CATNAME="Ohio"
+    $CATEGORIZE
+    rm Ohio.txt
+    unset OHIO
+  fi
+
+  if [ "$CLEVELAND" != "" ];
+  then
+    printf "%s" "$CLEVELAND" > Cleveland.txt
+    export CATFILE="Cleveland.txt"
+    export CATNAME="Cleveland"
+    $CATEGORIZE
+    rm Cleveland.txt
+    unset CLEVELAND
+  fi
+
+  if [ "$CINCINNATI" != "" ];
+  then
+    printf "%s" "$CINCINNATI" > Cincinnati.txt
+    export CATFILE="Cincinnati.txt"
+    export CATNAME="Cincinnati"
+    $CATEGORIZE
+    rm Cincinnati.txt
+    unset CINCINNATI
+  fi
+
+  debug_end "Ohio"
+
 fi
