+++ catscripts/Government/Countries/Venezuela/Venezuela.sh
@@ -2,14 +2,23 @@
 
 KEYWORDS_VENEZUELA="Venezuela"
 
-VENEZUELA=$(egrep -i "$KEYWORDS_VENEZUELA" newpages.txt)
-
-if [ "$VENEZUELA" != "" ];
+if [ "$1" == "" ];
 then
-  printf "%s" "$VENEZUELA" > Venezuela.txt
-  export CATFILE="Venezuela.txt"
-  export CATNAME="Venezuela"
-  $CATEGORIZE
-  rm Venezuela.txt
-  unset VENEZUELA
+
+  debug_start "Venezuela"
+
+  VENEZUELA=$(egrep -i "$KEYWORDS_VENEZUELA" newpages.txt)
+
+  if [ "$VENEZUELA" != "" ];
+  then
+    printf "%s" "$VENEZUELA" > Venezuela.txt
+    export CATFILE="Venezuela.txt"
+    export CATNAME="Venezuela"
+    $CATEGORIZE
+    rm Venezuela.txt
+    unset VENEZUELA
+  fi
+
+  debug_end "Venezuela"
+
 fi
